+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `cli/move`: Add option `-h` for help
 - `cli/config`: Add support for environment variable `HOME` in path
 - `scripts/install`: Add skeleton for installation script
+- `cli/list`: Add option to show project name header
 
 ### Changed